Centralize DXF write logic and improve type safety
Refactor EoDxfWrite to use centralized WriteCode* helpers for all DXF group code output, replacing direct m_writer calls. Add error tracking (m_writeOk) and stream state checks for robust write operations. Introduce GetBool in EoDxfHeader and ensure all boolean DXF codes (e.g., 290) are handled as bool, not int16. Update EoDxfEntity, EoDxfMLeader, and EoDxfDimensionStyle to use correct types for visibility and boolean fields. Improve code clarity, error handling, and DXF compliance throughout.

+++ Core Libraries/EoDxfLib/EoDxfWriter.cpp
@@ -131,6 +131,6 @@ bool EoDxfWriterAscii::WriteDouble(int code, double data) {
 
 // saved as int or add a bool member??
 bool EoDxfWriterAscii::WriteBool(int code, bool data) {
-  *m_fileStream << code << std::endl << data << std::endl;
+  *m_fileStream << code << std::endl << static_cast<int>(data) << std::endl;
   return (m_fileStream->good());
 }
